use the appropriate C++ 'cast' operator and fix the South plugin API signatures
The C cast operator for the 'PLUGIN_HANDLE' variable is too permissive.

Use the C++ 'static_cast': the cast operation is performed at the build
time and the compiler checks the compatibility and consistency between
the types.

We also need to fix the signatures of the API functions that contain
'PLUGIN_HANDLE',
and because of this, we are consistant with the North plugin API.

The Fledge documentation [1], that contains an example based on this
DHT11 South plugin, will also be updated.

[1] https://fledge-iot.readthedocs.io/en/develop/plugin_developers_guide/03_south_C_plugins.html

Signed-off-by: Mikael Bourhis <mikael.bourhis@smile.fr>